+++ src/WithoutRead/main.cpp
@@ -66,7 +66,7 @@ int main() {
             myTester.scan(queryCount, dataMin, dataMax, rangeSize, hhDataRange, hhProbability);
         }
     } while(line[0].compare(end) != 0);    
-    vector<unsigned int> runtimes = myTester.getRuntimes();
+    vector<unsigned long int> runtimes = myTester.getRuntimes();
     cout << "RUNTIMES:" << endl << runtimes[0];
     for (int i = 1; i < runtimes.size(); ++i){
         cout << "," << runtimes[i];
